+++ recipe/collabllm/collabllm_interation.py
@@ -21,9 +21,8 @@ from typing import Any, Optional
 from uuid import uuid4
 
 from recipe.collabllm.utils import remove_think_block
+from verl.interactions.base import BaseInteraction
 from verl.utils.rollout_trace import rollout_trace_op
-
-from .base import BaseInteraction
 
 logger = logging.getLogger(__name__)
 logger.setLevel(os.getenv("VERL_LOGGING_LEVEL", "WARN"))
