+++ recipe/collabllm/config/agent.yaml
@@ -1,0 +1,2 @@
+- name: collabllm_agent
+  _target_: recipe.collabllm.collabllm_agent_loop.CollabLLMAgentLoop
