+++ verl/workers/reward_manager/__init__.py
@@ -14,7 +14,6 @@
 
 from .registry import get_reward_manager_cls, register  # noqa: I001
 from .batch import BatchRewardManager
-from .collabllm import CollabLLMRewardManager
 from .dapo import DAPORewardManager
 from .naive import NaiveRewardManager
 from .prime import PrimeRewardManager
@@ -22,7 +21,6 @@ from .prime import PrimeRewardManager
 # Note(haibin.lin): no need to include all reward managers here in case of complicated dependencies
 __all__ = [
     "BatchRewardManager",
-    "CollabLLMRewardManager",
     "DAPORewardManager",
     "NaiveRewardManager",
     "PrimeRewardManager",
